SMS-4 Added convertCreditsToAmount and convertAmountToCredits methods and unit tests.

+++ sms_smpp/src/main/java/org/sakaiproject/sms/api/SmsBilling.java
@@ -124,9 +124,9 @@ public interface SmsBilling {
 	 * @param creditCount
 	 *            the credit count
 	 * 
-	 * @return the double
+	 * @return the credit amount
 	 */
-	public Double convertCreditsToAmount(int creditCount);
+	public Float convertCreditsToAmount(int creditCount);
 
 	/**
 	 * Convert amount to credits.
